+++ CHANGELOG.md
@@ -1,3 +1,44 @@
+## 0.7.0 (2020-09-22)
+
+### Features
+
+-   Added support for marine regions
+-   Refactor casts methods and types
+-   MRGID uses correct polygon instead of bounding box
+
+### Bug fixes
+
+-   Minor bugfixes
+
+## 0.6.4 (2020-09-11)
+
+### Features
+
+-   Added sequence metadata
+
+### Bug fixes
+
+-   Minor bugfixes
+
+## 0.6.2 (2020-09-09)
+
+### Features
+
+-   Added sequence metadata
+-   Added cast source file download
+
+### Bug fixes
+
+-   Minor bugfixes
+
+## 0.6.1 (2020-09-03)
+
+### Features
+
+### Bug fixes
+
+-   Minor bugfixes
+
 ## 0.3.1 (2020-03-24)
 
 ### Bug fixes